// src/objects/enums/EndMessages.ts
export enum EndMessages {
  English_Good = "You won from ",
  English_Bad = "The adversary has defeated you...",
  English_Giveup = "You have withdrawn.",
  English_Best = "You managed to be friends with ",
  French_Good = "Vous avez réussi à vaincre ",
  French_Bad = "L'adversaire vous a vaincu...",
  French_Giveup = "Vous avez déclaré forfait.",
  French_Best = "Vous avez réussi à être ami avec ",
  English_ExclamationMark = "!",
  French_ExclamationMark = " !",
}