// src/objects/enums/RetryMessages.ts
export enum RetryMessages {
  English_Good = "Will you get it better?",
  English_Bad = "Retry.",
  English_Giveup = "Maybe next time?",
  English_Best = "Can you be friends with other creatures?",
  French_Good = "Arriverez vous a faire mieux ?",
  French_Bad = "Rententez votre chance.",
  French_Giveup = "Peut être une prochaine fois ?",
  French_Best = "Pourrez vous être ami avec d'autres créatures ?",
}